Tic Tac Toe Solo Project
A solo project that uses vue.js to create a functioning game of tic-tac-toe. It may not be the most complex project possible, but it was a challenging experience for me. I used some tutorials from the official vue.js site to make the css transitions and the loop to check for wins is very similar to our connect four lab. Unfortunately I'm not a mega genius like some folk in our class so this was indeed a challenge for me.